services/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
tortoise-orm==0.20.0
aerich==0.7.2
asyncpg==0.27.0
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
dict2xml==1.7.4
pika==1.3.1
aio-pika==9.4.0
fastapi-mail==1.4.1
aiofiles==23.2.1